Further minor lock modification for speedup
Slightly irritating that this small modificaiton is the best I've 
managed to do for now as far as speeding things up goes. But the root of 
the problem appears to be the array that is being assigned to have some 
allocation/sizing issues in the related locked segment of code, this can 
lead to a slightly confusingd diganostic in a different location and/or 
asserts.