Fix race-condition esque problem
There was a problem where some values in the end result array would be 
indexed into multiple times and some wouldn't be indexed into at all.

This problem was introudced via a combination of the new thread pool and 
the way we were mapping arguments to threads. In essence each thread id 
has a specific argument tied to it, so it will only ever technically 
index into those arguments. However, the thread pool (especially with 
small examples) would complete tasks quick enough that it'd sometimes 
not use every single available thread, so it would skip some of the 
required arguments needed to compute things and also recompute the same 
arguments again (because it was running on a thread that was already 
used). 

To fix this I've disconnected the threadid from the argument map so the 
arguments are now thread agnostic, it's based on the EvalInfo a 
thread/task owns.