Fix race condition, make modifications to get several new functions working
- Moved lock to fix race condition
- Added ThreadArguementMap Copy Back on completion of computation, so 
that the iterator of the main EvalInfo is updated appropriately
- Reorganize the EvalStmtResult check code after the execution of the 
parallel components so that we can handle other scenarios, like new 
ESR_ParEarlySucceeded which is required when things early exit but do 
not feedback an ESR_Break or ESR_Returned.

The above changes allows several new algorithms to function in at least 
the capacity shown in the test cases.

The list of what's now working and what still hasn't been tested can be 
found in the function_test_list.txt file.