Compiler Executor Parallelism Check
* Addition of a check for an constexpr executor before enacting compile
time parallelism, it currently works by:
1) Checking if the executor is correct by checking the namespace its in
and it's name is correct
2) Checking if the function the executor is being passed to resides in
the std:: namespace

if both of the above are correct it'll try to parallelize the for loops
within the function body. The reason for the secondary check is because
we can't just parallelize any function that takes a constexpr executor.
What if a user wants to just pass a constexpr executor around? I think
the best method to implement this for the future would be to have an
attribute (keyword in the far off future?) that's applied to a funcitons
(SYCL style) that basically notates that it's a constexpr
parallelizeable function, this means users can make their own and we
don't have to restrict parallelizeable functions to the std:: namespace.

Other minor modifications:
* modified some loop iterator int's to size_t, to avoid warnings
* added some more function names to the function that chekcs whitelisted
function names, mainly to keep the list accurate for the time being, the
function isn't actually used at the moment