Implement set_intersection
This manages to get set_intersection to work at the moment, as a first
pass.

Although, I'm not convinced it's the right way to do it. I think the way
I've defined it within algorithm_constexpr_par may not be quite correct.

Although if I do define it correctly I think it'll open a whole new can
of worms that I plan to look into in the near future (essentially the
way its defined just now the result array itself is partitioned, when in
reality the algorithm calls for writing to the exact same index for each
thread and incrementing the index as its written to, this avoids the
need for a reduction but will likely cause a race condition).

For now this implementation works and is segmented to our code rather
than infecting other components. It may be getting a little compute
heavy, although I don't think its enough to offset the parallelism on
larger data sets though.