Adding new tests and script to run all
new functional tests are:
 copy_backward
 copy_n
 fill
 fill_n

all work except copy_backward, which I hope to get implemented soon.

Benchmarking scripts are:

set_intersection_par_vs_linear

which is testing the parallel versus linear implementation,
unfortunately the parallel implementation with locks is quite slow... so
I may have to revert this to the previous version (which may not work
perfectly in all cases) without locks. It could be interesting to have a
strict/vs non-strict toggle