Addition of new test cases, the majority work, a few are WIP
The majority of these tests work, some are WIP like includes, copy_if
and iter_swap. A good way to work out if something is (or should be)
working is to check the list of functions ran in the execute_tests.sh
any working tests are ran from the script to check for regressions.