Minor modification to old LoopWrappers and new addition
I modified some of the functions to capture variables by reference,
technically faster at run time if it's for some reason ever required.
However, it's mostly because it simplifies the AST in some cases eliding
some nodes that would be a pain to deal with. This problem may come up
again however.

New addition of __ImmutableVariable basically states that we don't want
to update/reduce a variable even if it's modified in one of the threads,
requires some follow up compiler code for it to actually function
though.