Several compiler additions, new features, cleanups and fixes
* Fix updating of return values for functions, previously we were
updating arguments but not return values. This should now be fixed.
* Add __ImmutableVariable compiler implementation, basically a black
list that's checked before we copy back from threads
* Some basic handling for Structures that are iterators containing a
pointer to some data. This isn't really a tractable problem for user
defined iterators unless we define some constraints or a specific name
users must use for the field/member that we should be partitioning
across the thread space.
* Some other minor cleanups, but a lot is still quite messy and needs
spring cleaning; for example an AST visitor that finds our ParamVar/Vars
as there could be more complex cases.
* A few other edge cases/bugs e.g. starting from offsets, not indexing
past the end of an array if we're using IteratorLoopStep and the
BeginEndIterator size is larger than the secondary data set (this still
requires a __last passed to the IteratorLoopStep, we can't guess the end
of a data set).
* Add some new nodes for visiting, sadly not all nodes are implicitly
visited which I had originally hoped for.

Other than the above really need to run clang format, the indentaition
is horrific at the moment as my IDE has no format option at the moment!