libcxx additions required for compilation of CEST
Mostly the addition of construct_at, bumping the C++ language standard
(not neccessarily required) up and adding constexpr to a bunch of things
CEST currently uses.

It's of note that the allocator's allocate and deallocate functions were
a little strange. They call a libcxx specific function that complains
about the same new and delete calls that I've added to allocate being
constexpr'd. However, those exact same functions work perfectly fine in
the const specialization of the allocator... a little bit of a head
scratcher.

Other than that it's worth while mentioning that the const version of
the allocator was already constexpr, but the non-const one was not.
However, Clang either has a bug or a feature that means if you have a
const constexpr value you still cannot modify it within a constexpr
call. So, the const specialization of the allocator would never work for
our purposes, the cppreference also doesn't distinguish between a const
and non-const specialization of the allocator it simply has the
allocator methods constexpr... in either case this current method works,
but I'm not sure it's spec compliant.