Add new Vector Tests, most of these are mildly modified variants of the std::array tests
25 out of 29 of the tests work at the moment, the final 4 tests all seem
to break due to one bug which causes an assertion that occurs when using
static constexpr vectors. I don't think this will be too hard to fix.

Other than that there is another bug I have managed to introduce with
the last patch or two that occurs when you try to push_back using the
size of a non-resized vector. This leads to an assert rather than any
useful compiler error or it passing, it might not be too difficult to
fix.