First Pass At Removing Lock
Still a significant number of problems with the tests at the moment.

This patch is in a middle state where things look a little worse than I
beleive they are, a fix for one problem is causing a lot of the tests to
break again, so I need to resolve the situation to achieve a decent pass
rate again.

However, the lock has been removed and things can run in parallel and
give the correct output in most cases.